backend/pyproject.toml
[project]
name = "random"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.4.18",
    "elasticsearch==8.11.0",
    "fastapi==0.104.1",
    "numpy==1.26.4",
    "pillow==10.1.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pypdf==3.17.1",
    "pytesseract==0.3.10",
    "python-docx==0.8.11",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    # 🔥 FIXED PART
    "sentence-transformers==2.2.2",
    "transformers==4.30.2",
    "huggingface-hub==0.14.1",
    "uvicorn==0.24.0",
    "google-genai==0.3.0",
]
